requirements.txt
playwright==1.56.0
langchain==1.1.0
langchain-openai==1.1.0
openai==2.8.1
python-dotenv==1.2.1
fastapi==0.122.0
uvicorn==0.38.0
pandas==2.3.3